refactor: reorganize admin routes and update import paths for consistency
